Replace the roles on the CLI by swap direction (buy-/sell-xmr)
The user should not be exposed to the cryptographic roles, thus we opted to remove them from the CLI.
We opted for using buy-xmr and sell-xmr for indicating the swap direction.
Eventually we could try to imply the role from currency params of send and receive.
This makes the CLI impl more complex though, because we have to cater for the allowed combinations.
It can be done in a later iteration.