Remove fragile assertions
The the Alice and Bob restart tests we were asserting that both
Alice's and Bob's balances were correct. This was problematic
because we are not waiting for both parties to complete causing
the test to fail (sometimes). This issue is resolved by removing
the assertions for the party that is moved into an async task as
we do not what state they have reached when we check their
balances.