Bitcoind RPC url, wallet name and Monero RPC url through config
On startup, if the config file does not exist yet initial setup is triggered and the user is queried for Bitcoind rpc url and Bitcoin wallet name, and Monero wallet rpc url.
The values entered by the user is written to config file in operating system specific default path.
Upon subsequent startups the value are read from the config file.

The corresponding CLI parameters were removed.