Optimize GitHub workflow file
1. Use rust-toolchain file toml syntax

By using the toml syntax, we can specify all required components
directly there and don't have to explicitly install the toolchain.

2. Use v2 of cache action

The newer version comes with zstd compression which allows for faster
saves and restores.

3. Run cache action in MacOS

The issue with MacOS having faulty caches has been fixed in a recent
release.

4. Bundle the registry cache together with the target directory

The v2 cache action can handle multiple paths.

5. Install tomlfmt based on cache hit

If we hit the cache, tomlfmt should be there. Otherwise we install it.