Make lock_xmr "idempotent"
Alice checks if xmr has already been locked (or in the process of being
locked) before locking xmr. This change means we no longer have to keep
track of whether we have locked xmr or not. We check the status of lock
xmr by looking for confirmed outgoing, pending or mempool transfers that
 match the shared public spend key and the amount.