Connect Alice to libp2p bootstrap nodes and bootstrap the DHT
This makes other nodes in the DHT aware of our public address which
- if the port is forwarded correctly - will be reachable from the
public internet.

For the `/dnsaddr` protocol to work, resolution of TXT records has
to succeed. At least on my machine, this didn't work out of the box.
Instead of every user having to mess with their local DNS settings,
we hardcode the quad9 DNS settings.