Search for Alice's connection details in the DHT
Instead of requiring the user to specify the connection details of
Alice, we automatically look them up in the DHT. Because we are now
connecting to more nodes than just Alice, a connection error is only
fatal if we fail a protocol with Alice. Failures with other nodes are
ignored.

If the user decides to directly specify connection details regardless,
we skip the DHT lookup and try to dial the peer directly.

In case the lookup in the DHT fails, we exit the eventloop.