Model event-loop state explicitely
This gives us an opportunity to add documentation to some of these
fields. This makes the whole design not a whole lot different from
what it was before with the tiny nuance that what is now called
"event-loop" is actually a loop and the state is called our more
explicitly.