Introduce read-only wallet for xpubs.
At the end of a swaps (may it be redeemed, refunded, or punished), the funds should be transferred to an address controlled by the according party.
This wallet should be decoupled from the protocol wallet, i.e. the wallet which will be used to sign internal transactions.
In this commit, we introduce a `personal_wallet` which is meant for exactly this.